fix: defend rubin-txgen portable key-file open
- keep Unix from-key-file secure open behind build tags

- fail closed for from-key-file on non-Unix platforms

- move Unix FIFO coverage out of all-platform tests

Layer: implementation/tests

Tests: go test ./cmd/rubin-txgen, cover, vet, git diff --check
