+++ clients/go/cmd/rubin-txgen/from_key_file_unsupported.go
@@ -1,0 +1,12 @@
+//go:build !unix
+
+package main
+
+import (
+	"errors"
+	"os"
+)
+
+func openRegularFromKeyFile(_ string) (*os.File, error) {
+	return nil, errors.New("from-key-file unsupported on this platform")
+}
