+++ clients/go/cmd/rubin-txgen/main.go
@@ -17,7 +17,6 @@ import (
 	"os"
 	"sort"
 	"strings"
-	"syscall"
 	"time"
 
 	"github.com/2tbmz9y2xt-lang/rubin-protocol/clients/go/consensus"
@@ -247,31 +246,6 @@ func loadFromKeyDER(fromKeyHex string, fromKeyFile string) ([]byte, error) {
 	default:
 		return decodeHexFlag(fromKeyHex)
 	}
-}
-
-func openRegularFromKeyFile(path string) (*os.File, error) {
-	fd, err := syscall.Open(path, syscall.O_RDONLY|syscall.O_CLOEXEC|syscall.O_NOFOLLOW|syscall.O_NONBLOCK, 0)
-	if err != nil {
-		if errors.Is(err, syscall.ELOOP) {
-			return nil, errors.New("from-key-file must be a regular file")
-		}
-		return nil, errors.New("read from-key-file failed")
-	}
-	f := os.NewFile(uintptr(fd), path)
-	if f == nil {
-		_ = syscall.Close(fd)
-		return nil, errors.New("read from-key-file failed")
-	}
-	openedInfo, err := f.Stat()
-	if err != nil {
-		_ = f.Close()
-		return nil, errors.New("read from-key-file failed")
-	}
-	if !openedInfo.Mode().IsRegular() {
-		_ = f.Close()
-		return nil, errors.New("from-key-file must be a regular file")
-	}
-	return f, nil
 }
 
 func readOpenedFromKeyFile(f *os.File) ([]byte, error) {
